+++ src/Simpletools/Db/Cassandra/Schema.php
@@ -4,6 +4,122 @@ namespace Simpletools\Db\Cassandra;
 
 class Schema
 {
+    protected $_tableName;
+    protected $_definition;
+
+    protected $_thisCreated = false;
+
+    public function __construct($name)
+    {
+        $this->_tableName = $name;
+    }
+
+    public function describe($definition)
+    {
+        $this->_definition = json_decode(json_encode($definition));
+
+        return $this;
+    }
+
+    public function create()
+    {
+        (new Client())->execute($this->toCql());
+
+        $this->_thisCreated = true;
+        return $this;
+    }
+
+    public function name($name=null)
+    {
+        if(!$name) return $this->_tableName;
+
+        $this->_tableName = $name;
+
+        return $this;
+    }
+
+    public function keyspace()
+    {
+        return $this->_definition->keyspace;
+    }
+
+    public function toCql()
+    {
+        $tableName = $this->_tableName;
+
+        $settings = $this->_definition;
+        $settings->keyspace = (new Client())->keyspace();
+
+        $columns = [];
+        $clusteringOrder = [];
+        foreach ($settings->columns as $column => $type)
+            $columns[] = "\"" . $column . "\" " . strtolower($type);
+        $primary = "PRIMARY KEY ((\"" . implode("\", \"", $settings->partition) . "\")";
+        if (isset($settings->clustering))
+        {
+            $clusteringKeys = [];
+            foreach ($settings->clustering as $key => $order)
+            {
+                $clusteringKeys[] = $key;
+                $clusteringOrder[] = "\"" . $key . "\" " . strtoupper($order);
+            }
+            $columns[] = $primary . ", \"" . implode("\", \"", $clusteringKeys) . "\")";
+        }
+        else
+        {
+            $columns[] = $primary . ")";
+        }
+
+        $_query = "CREATE TABLE IF NOT EXISTS " . $settings->keyspace . ".\"" . $tableName . "\" (\n";
+        $_query .= implode(", \n", $columns);
+        $_query .= ")";
+
+        if (count($clusteringOrder))
+        {
+            $_query .= " \nWITH CLUSTERING ORDER BY (" . implode(", ", $clusteringOrder) . ")";
+        }
+
+        $_parts = [];
+        if(isset($settings->options) && is_iterable($settings->options))
+        {
+            foreach ($settings->options as $option => $value) {
+                $_parts[] = "\"" . $option . "\" = " . $this->prepareOptionValue($value);
+            }
+
+            if (count($_parts)) {
+                if (count($clusteringOrder)) {
+                    $_query .= " \nAND";
+                } else {
+                    $_query .= " WITH";
+                }
+                $_query .= " " . implode(" \nAND ", $_parts);
+            }
+        }
+
+        return $_query;
+    }
+
+    /* HELPER METHODS */
+    public function prepareOptionValue($value)
+    {
+        if (is_object($value))
+        {
+            $value = json_encode($value);
+            $value = str_replace('\"', "%SQUOTE%", $value);
+            $value = str_replace("\"", "'", $value);
+            $value = str_replace("%SQUOTE%", '"', $value);
+        }
+        else if (is_array($value))
+        {
+            $value = json_encode($value);
+        }
+        else if (is_string($value))
+        {
+            $value = "'" . $value . "'";
+        }
+        return $value;
+    }
+
 	protected static $_schema = array();
 
 	public static function getSchema(Client $client, $keyspace, $table)
@@ -96,6 +212,4 @@ class Schema
 
 		return self::$_schema[$cluster][$keyspace][$table]['clusteringKey'];
 	}
-
-
 }